firebase.jsにconfigを記述
