+++ src/firebase.js
@@ -1,7 +1,15 @@
 import firebase from 'firebase/app';
 import 'firebase/auth';
 
-import FIREBASE_CONFIG from './env.js';
+const FIREBASE_CONFIG = {
+  apiKey: process.env.REACT_APP_API_KEY,
+  authDomain: process.env.REACT_APP_AUTH_DOMAIN,
+  databaseURL: process.env.REACT_APP_DATABASE_URL,
+  projectId: process.env.REACT_APP_PROJECT_ID,
+  storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
+  messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
+  appId: process.env.REACT_APP_APP_ID,
+};
 
 if (!firebase.apps.length) {
   firebase.initializeApp(FIREBASE_CONFIG);
